refactor: scan helpers, handler split, backfill consolidation, TestMain cleanup
- Add column constants and scan helpers (decisionCols/scanOneDecision,
  agentCols/scanOneAgent, apiKeyCols/scanOneAPIKey, grantCols/scanOneGrant)
  to eliminate duplicated inline scan patterns across storage layer

- Extract queueSearchOutbox helper and txExecer interface to share
  outbox insertion logic between decisions.go and trace.go

- Split handlers_decisions.go (1525 lines) into handlers_conflicts.go
  and handlers_assessments.go by domain, with shared parsePathUUID and
  parseConflictFilters helpers

- Consolidate BackfillEmbeddings and BackfillOutcomeEmbeddings into
  shared backfillBatch method via backfillSpec struct

- Remove deprecated QualityScore field from Decision model

- Consolidate 4 TestMain functions to use testutil.MustStartTimescaleDB(),
  replacing ~200 lines of duplicated testcontainer boilerplate

All tests pass with -race -count=1. No behavioral changes.
