+++ internal/storage/grants.go
@@ -14,6 +14,33 @@ import (
 	"github.com/ashita-ai/akashi/internal/model"
 )
 
+// grantCols is the SELECT column list for the standard 9-column grant query.
+const grantCols = `id, org_id, grantor_id, grantee_id, resource_type, resource_id, permission, granted_at, expires_at`
+
+// scanOneGrant scans the 9-column grantCols from a single row.
+func scanOneGrant(row pgxRowScanner) (model.AccessGrant, error) {
+	var g model.AccessGrant
+	if err := row.Scan(
+		&g.ID, &g.OrgID, &g.GrantorID, &g.GranteeID, &g.ResourceType, &g.ResourceID,
+		&g.Permission, &g.GrantedAt, &g.ExpiresAt,
+	); err != nil {
+		return model.AccessGrant{}, fmt.Errorf("storage: scan grant: %w", err)
+	}
+	return g, nil
+}
+
+func scanGrants(rows pgx.Rows) ([]model.AccessGrant, error) {
+	var grants []model.AccessGrant
+	for rows.Next() {
+		g, err := scanOneGrant(rows)
+		if err != nil {
+			return nil, err
+		}
+		grants = append(grants, g)
+	}
+	return grants, rows.Err()
+}
+
 // CreateGrant inserts a new access grant.
 func (db *DB) CreateGrant(ctx context.Context, grant model.AccessGrant) (model.AccessGrant, error) {
 	if grant.ID == uuid.Nil {
@@ -119,15 +146,10 @@ func (db *DB) DeleteGrantWithAudit(ctx context.Context, orgID, id uuid.UUID, aud
 // tenant isolation. Even though the handler performs its own authz check,
 // the storage layer must enforce org boundaries on every query.
 func (db *DB) GetGrant(ctx context.Context, orgID uuid.UUID, id uuid.UUID) (model.AccessGrant, error) {
-	var g model.AccessGrant
-	err := db.pool.QueryRow(ctx,
-		`SELECT id, org_id, grantor_id, grantee_id, resource_type, resource_id,
-		 permission, granted_at, expires_at
-		 FROM access_grants WHERE id = $1 AND org_id = $2`, id, orgID,
-	).Scan(
-		&g.ID, &g.OrgID, &g.GrantorID, &g.GranteeID, &g.ResourceType, &g.ResourceID,
-		&g.Permission, &g.GrantedAt, &g.ExpiresAt,
+	row := db.pool.QueryRow(ctx,
+		`SELECT `+grantCols+` FROM access_grants WHERE id = $1 AND org_id = $2`, id, orgID,
 	)
+	g, err := scanOneGrant(row)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return model.AccessGrant{}, fmt.Errorf("storage: grant %s: %w", id, ErrNotFound)
@@ -199,8 +221,7 @@ func (db *DB) ListGrants(ctx context.Context, orgID uuid.UUID, limit, offset int
 	}
 
 	rows, err := db.pool.Query(ctx,
-		`SELECT id, org_id, grantor_id, grantee_id, resource_type, resource_id,
-		 permission, granted_at, expires_at
+		`SELECT `+grantCols+`
 		 FROM access_grants
 		 WHERE org_id = $1
 		 ORDER BY granted_at DESC
@@ -211,25 +232,17 @@ func (db *DB) ListGrants(ctx context.Context, orgID uuid.UUID, limit, offset int
 	}
 	defer rows.Close()
 
-	var grants []model.AccessGrant
-	for rows.Next() {
-		var g model.AccessGrant
-		if err := rows.Scan(
-			&g.ID, &g.OrgID, &g.GrantorID, &g.GranteeID, &g.ResourceType, &g.ResourceID,
-			&g.Permission, &g.GrantedAt, &g.ExpiresAt,
-		); err != nil {
-			return nil, 0, fmt.Errorf("storage: scan grant: %w", err)
-		}
-		grants = append(grants, g)
+	grants, err := scanGrants(rows)
+	if err != nil {
+		return nil, 0, err
 	}
-	return grants, total, rows.Err()
+	return grants, total, nil
 }
 
 // ListGrantsByGrantee returns all active grants for a grantee within an org.
 func (db *DB) ListGrantsByGrantee(ctx context.Context, orgID uuid.UUID, granteeID uuid.UUID) ([]model.AccessGrant, error) {
 	rows, err := db.pool.Query(ctx,
-		`SELECT id, org_id, grantor_id, grantee_id, resource_type, resource_id,
-		 permission, granted_at, expires_at
+		`SELECT `+grantCols+`
 		 FROM access_grants
 		 WHERE org_id = $1 AND grantee_id = $2 AND (expires_at IS NULL OR expires_at > now())
 		 ORDER BY granted_at DESC`, orgID, granteeID,
@@ -239,16 +252,5 @@ func (db *DB) ListGrantsByGrantee(ctx context.Context, orgID uuid.UUID, granteeI
 	}
 	defer rows.Close()
 
-	var grants []model.AccessGrant
-	for rows.Next() {
-		var g model.AccessGrant
-		if err := rows.Scan(
-			&g.ID, &g.OrgID, &g.GrantorID, &g.GranteeID, &g.ResourceType, &g.ResourceID,
-			&g.Permission, &g.GrantedAt, &g.ExpiresAt,
-		); err != nil {
-			return nil, fmt.Errorf("storage: scan grant: %w", err)
-		}
-		grants = append(grants, g)
-	}
-	return grants, rows.Err()
+	return scanGrants(rows)
 }
